fix data fetching in profile
