+++ front/src/components/profile/DesktopProfile.tsx
@@ -3,12 +3,45 @@ import { Navbar } from '@/components/profile/Navbar';
 import { Statis } from '@/components/profile/UserStats';
 import MyAvatar from './MyAvatar';
 import { userType } from '@/types/types';
+import { useEffect, useState } from 'react';
+import Cookies from 'js-cookie';
+import axios from 'axios';
 
 interface Props {
   user: userType | null;
 }
 
-export const DesktopProfile: React.FC<Props> = ({ user }): JSX.Element => {
+function getUserData(): userType | null {
+  const [user, setUser] = useState<userType | null>(null);
+
+  useEffect(() => {
+    const jwtToken = Cookies.get('token');
+    const userId = Cookies.get('_id');
+    axios
+      .get<userType | null>(
+        `${process.env.NEXT_PUBLIC_API_URL}/users/${userId}`,
+        {
+          headers: {
+            Authorization: `Bearer ${jwtToken}`,
+            'Content-Type': 'application/json',
+          },
+        }
+      )
+      .then((response) => {
+        if (response.status === 200) {
+          setUser(response.data);
+        }
+      })
+      .catch((error) => {
+        console.log(error);
+        return null;
+      });
+  }, []);
+  return user;
+}
+
+export const DesktopProfile = (): JSX.Element => {
+  const userData = getUserData();
   return (
     <div className="flex flex-col gap-10 mx-[20px]">
       <div className="bg-[#243230] relative rounded-[20px] max-h-[333px]">
@@ -20,7 +53,7 @@ export const DesktopProfile: React.FC<Props> = ({ user }): JSX.Element => {
               className={`bg-secondary flex justify-center items-center
                       rounded-full text-[#161F1E] text-[1.25em] font-bold  h-[40px] w-[40px]`}
             >
-              {user?.Level}
+              {userData?.Level}
             </span>
           </div>
         </div>
@@ -33,19 +66,19 @@ export const DesktopProfile: React.FC<Props> = ({ user }): JSX.Element => {
             <div className="flex items-center justify-between py-8 w-1/3">
               <Statis
                 title="Total game"
-                total={user?.TotalMatch ? user?.TotalMatch : 0}
+                total={userData?.TotalMatch ? userData?.TotalMatch : 0}
                 line={true}
                 style="text-[1.5em] md:text-[1.1em]"
               />
               <Statis
                 title="Wins"
-                total={user?.TotalWin ? user?.TotalWin : 0}
+                total={userData?.TotalWin ? userData?.TotalWin : 0}
                 line={true}
                 style="text-[1.5em] md:text-[1.1em]"
               />
               <Statis
                 title="Loses"
-                total={user?.TotalLose ? user?.TotalLose : 0}
+                total={userData?.TotalLose ? userData?.TotalLose : 0}
                 line={false}
                 style="text-[1.5em] md:text-[1.1em]"
               />
@@ -53,7 +86,7 @@ export const DesktopProfile: React.FC<Props> = ({ user }): JSX.Element => {
 
             <div className="w-1/3 flex justify-center items-center">
               <h2 className="text-[2em] md:text-[1.5em] text-white font-bold mt-12">
-                {user?.login}
+                {userData?.login}
               </h2>
             </div>
 
@@ -64,7 +97,7 @@ export const DesktopProfile: React.FC<Props> = ({ user }): JSX.Element => {
                 </h3>
                 <div className="flex justify-center items-center">
                   <h3 className="uppercase text-white text-[1.5m] md:text-[1.1em] font-bold pl-2">
-                    {user?.location ? user?.location : 'Unvailable'}
+                    {userData?.location ? userData?.location : 'Unvailable'}
                   </h3>
                 </div>
               </div>
